+++ tests/SixToFix.Web.Tests/GlobalUsings.cs
@@ -15,6 +15,4 @@ global using SixToFix.Domain.Entities;
 global using SixToFix.Web.Components.Shared;
 global using SixToFix.Web.Navigation;
 global using SixToFix.Web.Pages;
-global using SixToFix.Web.Realtime;
-global using SixToFix.Web.Tests.Fakes;
 global using Xunit;